Publish script

+++ publish.sh
@@ -62,7 +62,15 @@ fi
 
 echo "========> Module: $modulename (key :  $repositorykey)"
 echo "Sending file to repository.."
-curl -F "file=@$modulename.zip" -F "preview=@$previewfile" -F "config=@$configinstall" -F "module=$modulename" -F "key=$repositorykey" $repositoryurl
+response=$(curl -F "file=@$modulename.zip" -F "preview=@$previewfile" -F "config=@$configinstall" -F "module=$modulename" -F "key=$repositorykey" $repositoryurl)
+echo $response
+
+if [[ $response = *"[ERROR]"* ]]; then
+    echo "[ERROR] has occurred"
+    exit 1
+fi
+
+echo "[SUCCESS] upload successfully!"
 
 rm $modulename.zip
 exit 0
